refactor: move demo editor Error to new file
